NODE-665: overloading translating binOp to functioncall
