+++ lang/jvm/src/test/scala/com/wavesplatform/lang/IntegrationTest.scala
@@ -16,10 +16,21 @@ class IntegrationTest extends PropSpec with PropertyChecks with Matchers with No
   }
 
   property("function call") {
-    eval[Long]("10 + 2".stripMargin) shouldBe Right(12)
+    eval[Long]("10 + 2") shouldBe Right(12)
   }
 
-  property("equals on byte array") {
-    eval[Boolean]("base58'3My3KZgFQ3CrVHgz6vGRt8687sH4oAA1qp8' == base58'3My3KZgFQ3CrVHgz6vGRt8687sH4oAA1qp8'".stripMargin) shouldBe Right(true)
+  property("equals works on primitive types") {
+//    eval[Boolean]("base58'3My3KZgFQ3CrVHgz6vGRt8687sH4oAA1qp8' == base58'3My3KZgFQ3CrVHgz6vGRt8687sH4oAA1qp8'") shouldBe Right(true)
+    eval[Boolean]("1 == 2") shouldBe Right(false)
+//    eval[Boolean]("3 == 3") shouldBe Right(true)
+//    eval[Boolean]("false == false") shouldBe Right(true)
+//    eval[Boolean]("true == false") shouldBe Right(false)
+//    eval[Boolean]("true == true") shouldBe Right(true)
+//    eval[Boolean]("""   "x" == "x"     """) shouldBe Right(true)
+//    eval[Boolean]("""   "x" == "y"     """) shouldBe Right(false)
+  }
+
+  property("equals shouldn't compile on option") {
+    eval[Boolean]("Some(1) == Some(2)") should produce("ERROR")
   }
 }
